+++ Editor/PackageData.cs
@@ -1,7 +1,40 @@
 ﻿using System;
+using System.Collections.Generic;
+using Newtonsoft.Json;
 
 namespace ParkMinPackages.PackageManager.Editor
 {
+	[Serializable]
+	internal class GitDependency
+	{
+		[JsonProperty("packageName")] public string PackageName;
+		[JsonProperty("url")] public string URL;
+	}
+
+	[Serializable]
+	internal class NuGetDependency
+	{
+		[JsonProperty("packageName")] public string PackageName;
+		[JsonProperty("version")] public string Version;
+	}
+
+	internal enum PackageDependencyState
+	{
+		Installed,
+		NotInstalled,
+		VersionMismatch,
+		Unavailable
+	}
+
+	internal struct PackageDependencyData
+	{
+		public string Name;
+		public string Version;
+		public string URL;
+		public string InstalledVersion;
+		public PackageDependencyState State;
+	}
+
 	internal struct PackageData
 	{
 		public string RepoName;
@@ -11,6 +44,8 @@ namespace ParkMinPackages.PackageManager.Editor
 		public string RemoteCommitHash;
 		public string CurrentCommitHash;
 		public bool IsEmbed;
+		public IReadOnlyList<PackageDependencyData> GitDependencies;
+		public IReadOnlyList<PackageDependencyData> NuGetDependencies;
 		public PackageState State
 		{
 			get